Splitting test files
