+++ tests/test_committee.py
@@ -1,27 +1,6 @@
 import unittest
-from pyopenfec import Candidate, Committee
+from pyopenfec import Committee
 from datetime import datetime
-
-
-class CandidateTest(unittest.TestCase):
-    def setUp(self):
-        self.candidate = None
-        candidates = Candidate.fetch(candidate_id="H8CA05035")
-        for candidate in candidates:
-            self.candidate = candidate
-
-    def test_candidate_dates(self):
-        self.assertIsInstance(self.candidate.first_file_date, datetime)
-        self.assertIsInstance(self.candidate.last_f2_date, datetime)
-        self.assertIsInstance(self.candidate.last_file_date, datetime)
-        self.assertIsInstance(self.candidate.load_date, datetime)
-
-    def test_history_dates(self):
-        history = self.candidate.history[2018]
-        self.assertIsInstance(history.first_file_date, datetime)
-        self.assertIsInstance(history.last_f2_date, datetime)
-        self.assertIsInstance(history.last_file_date, datetime)
-        self.assertIsInstance(history.load_date, datetime)
 
 
 class CommitteeTest(unittest.TestCase):
@@ -77,7 +56,3 @@ class CommitteeTest(unittest.TestCase):
             example = donation
         self.assertIsInstance(example.disbursement_date, datetime)
         self.assertIsInstance(example.load_date, datetime)
-
-
-if __name__ == "__main__":
-    unittest.main()
